requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.3
cryptography==3.2.1
decorator==4.4.2
psycopg2==2.8.6
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.4
six==1.15.0
tornado==6.1
tzlocal==2.1
